requirements.txt
pandas
matplotlib
seaborn
scikit-learn
pythainlp
requests
beautifulsoup4
python-dotenv